requirements.txt
google-generativeai==0.7.2
requests==2.32.3
beautifulsoup4==4.12.3
python-dotenv==1.0.1
pydantic==2.9.2

